% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{compute_circumference}
\alias{compute_circumference}
\title{Compute Circumference of the smallest circle that could circumscribe the flower}
\usage{
compute_circumference(dataframe)
}
\arguments{
\item{dataframe}{data.frame with columns Sepal.Length and Petal.Length}
}
\value{
data.frame with new column Circumference
}
\description{
using the larger length as the radius
}
\examples{
compute_circumference(iris)
}
